add Game, Player, Human Player, Input classes
In the Game class there will be methods for game flow (plus checks for end of the game).
Player class is an abstract class, the game will have two classes that extend this class (Human and AI classes).
HumanPlayer class extends Player and implements its abstract method for placing a piece on the board (if there is an empty spot).
Input class is a singleton class and its getInt() method is designed to return an int value;
It is useful in case that there are two human players that play one against the other, so we don't have to open and close the scanner object after each move.