add AI_Player class, make Game class singleton
Game class is made as a singleton in order to be able to acces the GameBoard object and isX boolean (used to keep the track of the players).
AI_Player extends Player class and has move() method implemented. It process all the possibilities and choose the most advantageous for the computer (win>drow>lose).
GameBoard class has the win method changed. Also, this class has a toString() method, intead of print() method.